# test-results/2025-09-03/crawler-tests/data/openapi_schema_170814.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "历史文本项目 - 数据源服务",
    "description": "\n    ## 数据源服务 API\n    \n    提供多平台内容爬取、代理管理和内容处理功能的微服务。\n    \n    ### 主要功能\n    - 🕷️ **爬虫管理**: 支持今日头条、百家号、小红书等平台\n    - 📄 **内容管理**: 手动添加、批量导入、查询和更新内容\n    - 🌐 **代理管理**: 代理池管理、自动测试和轮换\n    - 📊 **监控统计**: 实时状态监控和数据统计\n    \n    ### 技术栈\n    - **框架**: FastAPI + Python 3.9+\n    - **数据库**: MongoDB + Redis\n    - **爬虫**: Scrapy + Selenium\n    - **代理**: 多供应商代理池\n    \n    ### 使用说明\n    1. 使用 `/crawlers/` 接口管理爬虫任务\n    2. 使用 `/content/` 接口管理内容数据\n    3. 使用 `/proxy/` 接口管理代理设置\n    4. 查看 `/health` 检查服务状态\n    ",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/crawlers/": {
      "post": {
        "tags": [
          "爬虫管理"
        ],
        "summary": "创建爬虫任务",
        "description": "创建新的爬虫任务\n\n- **platform**: 平台类型 (toutiao, baijiahao, xiaohongshu)\n- **keywords**: 关键词列表，用于内容抓取\n- **max_pages**: 最大抓取页数\n- **interval**: 请求间隔时间，避免过于频繁请求\n- **priority**: 任务优先级，影响执行顺序",
        "operationId": "create_crawler_task_api_v1_crawlers__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCrawlerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Crawler Task Api V1 Crawlers  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "爬虫管理"
        ],
        "summary": "获取所有任务",
        "description": "获取所有爬虫任务列表，支持分页和过滤",
        "operationId": "get_all_crawler_tasks_api_v1_crawlers__get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CrawlerStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按状态过滤",
              "title": "Status"
            },
            "description": "按状态过滤"
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContentSource"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按平台过滤",
              "title": "Platform"
            },
            "description": "按平台过滤"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Size"
            },
            "description": "每页数量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get All Crawler Tasks Api V1 Crawlers  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/content/": {
      "post": {
        "tags": [
          "内容管理"
        ],
        "summary": "添加单个内容",
        "description": "手动添加单个内容\n\n- **title**: 内容标题\n- **content**: 内容正文\n- **source**: 内容来源\n- **author**: 作者（可选）\n- **source_url**: 原始链接（可选）\n- **publish_time**: 发布时间（可选）\n- **content_type**: 内容类型（默认为文章）",
        "operationId": "create_content_api_v1_content__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Content Api V1 Content  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "内容管理"
        ],
        "summary": "获取内容列表",
        "description": "获取内容列表，支持多种过滤和排序选项",
        "operationId": "get_content_list_api_v1_content__get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按状态过滤",
              "title": "Status"
            },
            "description": "按状态过滤"
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContentSource"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按来源过滤",
              "title": "Source"
            },
            "description": "按来源过滤"
          },
          {
            "name": "content_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContentType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按类型过滤",
              "title": "Content Type"
            },
            "description": "按类型过滤"
          },
          {
            "name": "author",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按作者过滤",
              "title": "Author"
            },
            "description": "按作者过滤"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按分类过滤",
              "title": "Category"
            },
            "description": "按分类过滤"
          },
          {
            "name": "keywords",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "关键词搜索（逗号分隔）",
              "title": "Keywords"
            },
            "description": "关键词搜索（逗号分隔）"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "开始日期 (YYYY-MM-DD)",
              "title": "Start Date"
            },
            "description": "开始日期 (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "结束日期 (YYYY-MM-DD)",
              "title": "End Date"
            },
            "description": "结束日期 (YYYY-MM-DD)"
          },
          {
            "name": "min_quality_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100.0,
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "最小质量分数",
              "title": "Min Quality Score"
            },
            "description": "最小质量分数"
          },
          {
            "name": "max_quality_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100.0,
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "最大质量分数",
              "title": "Max Quality Score"
            },
            "description": "最大质量分数"
          },
          {
            "name": "min_view_count",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "最小浏览量",
              "title": "Min View Count"
            },
            "description": "最小浏览量"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Size"
            },
            "description": "每页数量"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "排序字段",
              "default": "created_at",
              "title": "Sort By"
            },
            "description": "排序字段"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "description": "排序顺序",
              "default": "desc",
              "title": "Sort Order"
            },
            "description": "排序顺序"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Content List Api V1 Content  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/proxy/": {
      "get": {
        "tags": [
          "代理管理"
        ],
        "summary": "获取代理列表",
        "description": "获取代理列表，支持过滤和分页",
        "operationId": "get_proxy_list_api_v1_proxy__get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProxyStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按状态过滤",
              "title": "Status"
            },
            "description": "按状态过滤"
          },
          {
            "name": "quality",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProxyQuality"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按质量过滤",
              "title": "Quality"
            },
            "description": "按质量过滤"
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按供应商过滤",
              "title": "Provider"
            },
            "description": "按供应商过滤"
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按国家过滤",
              "title": "Country"
            },
            "description": "按国家过滤"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Size"
            },
            "description": "每页数量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Proxy List Api V1 Proxy  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentSource": {
        "type": "string",
        "enum": [
          "toutiao",
          "baijiahao",
          "xiaohongshu",
          "manual",
          "rss",
          "api"
        ],
        "title": "ContentSource",
        "description": "内容来源枚举"
      },
      "ContentStatus": {
        "type": "string",
        "enum": [
          "pending",
          "processing",
          "completed",
          "failed",
          "rejected"
        ],
        "title": "ContentStatus",
        "description": "内容状态枚举"
      },
      "ContentType": {
        "type": "string",
        "enum": [
          "article",
          "video",
          "image",
          "audio",
          "other"
        ],
        "title": "ContentType",
        "description": "内容类型枚举"
      },
...
      "CrawlerStatus": {
        "type": "string",
        "enum": [
          "idle",
          "running",
          "paused",
          "stopped",
          "error",
          "finished"
        ],
        "title": "CrawlerStatus",
        "description": "爬虫状态枚举"
      },
      "CreateCrawlerRequest": {
        "properties": {
          "platform": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ContentSource"
              }
            ],
            "description": "平台类型"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Keywords",
            "description": "关键词列表"
          },
          "max_pages": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Pages",
            "description": "最大抓取页数",
            "default": 10
          },
          "interval": {
            "type": "number",
            "maximum": 60.0,
            "minimum": 0.1,
            "title": "Interval",
            "description": "请求间隔(秒)",
            "default": 5.0
          },
          "timeout": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 30.0,
            "title": "Timeout",
            "description": "超时时间(秒)",
            "default": 300
          },
          "priority": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TaskPriority"
              }
            ],
            "description": "任务优先级",
            "default": 5
          },
          "proxy_enabled": {
            "type": "boolean",
            "title": "Proxy Enabled",
            "description": "是否启用代理",
            "default": true
          },
          "retry_attempts": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Retry Attempts",
            "description": "重试次数",
            "default": 3
          },
          "custom_settings": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Settings",
            "description": "自定义配置"
          }
        },
        "type": "object",
        "required": [
          "platform",
          "keywords"
        ],
        "title": "CreateCrawlerRequest",
        "description": "创建爬虫任务请求模型",
        "example": {
          "custom_settings": {},
          "interval": 5.0,
          "keywords": [
            "历史",
            "文化"
          ],
          "max_pages": 10,
          "platform": "toutiao",
          "priority": 5,
          "proxy_enabled": true,
          "retry_attempts": 3,
          "timeout": 300
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProxyQuality": {
        "type": "string",
        "enum": [
          "unknown",
          "low",
          "medium",
          "high",
          "premium"
        ],
        "title": "ProxyQuality",
        "description": "代理质量等级"
      },
      "ProxyStatus": {
        "type": "string",
        "enum": [
          "active",
          "inactive",
          "testing",
          "banned",
          "timeout",
          "error"
        ],
        "title": "ProxyStatus",
        "description": "代理状态枚举"
      },
...
      "TaskPriority": {
        "type": "integer",
        "enum": [
          1,
          5,
          8,
          10
        ],
        "title": "TaskPriority",
        "description": "任务优先级枚举"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}